Refactor: Move server logic from main.py to src/server.py.
