+++ src/templates/json/stock_data.json
