+++ src/templates/margin_of_safety.html
