+++ src/templates/txt/management_tooltip.txt
