+++ src/templates/txt/margin_of_safety_tooltip.txt
