+++ src/templates/txt/market_cap_tooltip.txt
