+++ src/templates/txt/meaning_tooltip.txt
