+++ src/templates/txt/moat_tooltip.txt
