+++ src/templates/txt/payback_time_tooltip.txt
